+++ sugar/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 2.1.0 (30/06/2021)
+
+This release adds new annotations & cryptological functions.
+
+- Add `@PlatformDependent`
+- Add `@Subset`
+- Add `@Throws`
+- Add `@annotation`
+- Add `@throws`
+- Add `HashExtensions`
+- Add `nonce(...)`
+- Change `Result.present` to `Result.successful`
+- Change `Result.notPresent` to `Result.failure`
+- Change `Result.ifPresent(...)` tp `Result.ifSuccessful(...)`
+- Change `Result.ifNotPresent(...)` tp `Result.ifFailure(...)`
+
 ## 2.0.1 - (29/06/2021)
 
 Downgrade `meta` from `1.4.0` to `.1.3.0` to enable compatibility with Flutter